+++ src/web/src/components/PortForwarding/PortForwarding.jsx
@@ -24,6 +24,7 @@ import {
 } from 'semantic-ui-react';
 
 const initialState = {
+  activeTab: 0,
   availablePorts: [],
   createForm: {
     destinationHost: '',
@@ -599,7 +600,11 @@ class PortForwarding extends Component {
         )}
 
         <Tab
+          activeIndex={this.state.activeTab}
           menu={{ pointing: true, secondary: true }}
+          onTabChange={(_event, { activeIndex }) =>
+            this.setState({ activeTab: activeIndex })
+          }
           panes={panes}
           renderActiveOnly={false}
           style={{ marginTop: '20px' }}